codebase-query-contextualiser/embedding-vector/requirements.txt
transformers
torch
sentence-transformers
qdrant-client
camel-ai
pillow
pandas
unstructured